+++ Core/MultilayerNetworkSerializer.cs
@@ -71,7 +71,7 @@ namespace Networks.Core
             // TODO: turn this into a proper state machine for robustness
             string line = string.Empty;
             MultilayerNetwork retVal = null;
-            List<Tuple<NodeTensor, NodeTensor, double>> interEdges = new List<Tuple<NodeTensor, NodeTensor, double>>();
+            List<Tuple<NodeTensor, NodeTensor, float>> interEdges = new List<Tuple<NodeTensor, NodeTensor, float>>();
 
             line = MoveToNextBlock(stIn);
             uint block; 
@@ -130,8 +130,8 @@ namespace Networks.Core
                                 NodeTensor tensorFrom = new NodeTensor(tensorParms[0], tensorParms[1]);
                                 tensorParms = edgeParams[1].Split(':');
                                 NodeTensor tensorTo = new NodeTensor(tensorParms[0], tensorParms[1]);
-                                double wt = Convert.ToDouble(edgeParams[2]);
-                                Tuple<NodeTensor, NodeTensor, double> tpl = new Tuple<NodeTensor, NodeTensor, double>(tensorFrom, tensorTo, wt);
+                                float wt = Convert.ToSingle(edgeParams[2]);
+                                Tuple<NodeTensor, NodeTensor, float> tpl = new Tuple<NodeTensor, NodeTensor, float>(tensorFrom, tensorTo, wt);
 
                                 // If the target elementary layer exists, add the edge, otherwise, add to the list to be resolved later
                                 if (retVal.HasElementaryLayer(tensorTo.coordinates))
@@ -151,7 +151,7 @@ namespace Networks.Core
             // now add the accumulated interlayer edges
             // TODO: add an opportunistic test after each elementary layer is complete so as to keep the length of this list down
             // for multilayer networks with lots of interlayer edges -- perf test to find a balance of entries in list and number of elementary layers instantiated
-            foreach (Tuple<NodeTensor, NodeTensor, double> tuple in interEdges)
+            foreach (Tuple<NodeTensor, NodeTensor, float> tuple in interEdges)
                 retVal.AddEdge(tuple.Item1, tuple.Item2, tuple.Item3);
             return retVal;
         }
